dom/media/test/test_play_promise_3.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Media test: promise-based play() method</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <script type="text/javascript" src="manifest.js"></script>
</head>
<body>
<pre id="test">

<script>
// Name: playAfterPlaybackStarted
// Case: invoke play() on an element that is already playing
// Expected result: resolve the promise

let manager = new MediaTestManager;

function initTest(test, token) {
  manager.started(token);

  let element = document.createElement(getMajorMimeType(test.type));
  element.preload = "auto";
  element.src = test.name;
  once(element, "playing").then(() => {
    ok(element.readyState >= HTMLMediaElement.HAVE_FUTURE_DATA, `playAfterPlaybackStarted(${token})`);
    ok(!element.paused, `playAfterPlaybackStarted(${token})`);
    element.play().then(
      (result) => {
        if (result == undefined) {
          ok(true, `${token} is resolved with ${result}.`);
        } else {
          ok(false, `${token} is resolved with ${result}.`);
        }
      },
      (error) => {
        ok(false, `${token} is rejected with ${error.name}.`);
      }
    ).then( () => { manager.finished(token); } );
  });

  element.play();
}

manager.runTests(gSmallTests, initTest);

</script>
